removed point2D properties and replaced them with Coordinates.

+++ src/main/java/com/savesystem/ArmorData.java
@@ -3,6 +3,6 @@ package com.savesystem;
 public class ArmorData {
     public String Name;
     public int defence;
-    public double x;
-    public double y;
+    public int x;
+    public int y;
 }
